# swagger.yaml
swagger: "2.0"
info:
  description: "This is the swagger doc for vanDevTimes"
  version: "1.0.0"
  title: "Swagger vanDevTimes"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "vantimes.info@gmail.com"
basePath: "/api"
paths:
...
  /scheduler:
    get:
      tags:
        - "Scheduler"
      summary: "Get scheduler settings of all tasks"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            example: {
              "scheduler": [
              {
                "ID": "approval-reminder",
                "FREQUENCY": 7,
                "START_TIME": "2019-01-01T08:00:00.000Z",
                "MAX_NUM": 17
              },
              {
                "ID": "RSS-retrieval",
                "FREQUENCY": 10,
                "START_TIME": "2019-01-01T08:00:00.000Z",
                "MAX_NUM": 17
              },
              {
                "ID": "subscription-email",
                "FREQUENCY": 3,
                "START_TIME": "2019-01-01T08:00:00.000Z",
                "MAX_NUM": 17
              }
              ]
            }
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          schema:
            example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
  /user:
    get:
      tags:
        - "User"
      summary: "get user account info"
      consumes:
        - "application/json"
      parameters:
        -  in: "body"
           name: "body"
           description: "a user email"
           schema:
             example: {
               "email": "regularUser@email.com"
             }
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          schema:
            example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
  /users:
    get:
      tags:
        - "User"
      summary: "get all users"

      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ArrayOfUsers"
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          schema:
            example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
  /user/permissions:
    put:
      tags:
        - "User"
        - "Permission"
      summary: "update a user's permission level"
      consumes:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "a user email and new role"
          schema:
            example: {
              "newRole": 2,
              "email": "regularUser@email.com"
            }
      responses:
        200:
          description: "successful operation"
          schema:
            example: {
              "newRole": 2,
              "email": "regularUser@email.com"
            }
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          schema:
            example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
  /company:
    put:
      tags:
        - "Company"
      summary: "set a user's company"
      consumes:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "a user email and new company"
          schema:
            example: {
              "newComp": "ABC",
              "email": "regularUser@email.com"
            }
      responses:
        200:
          description: "successful operation"
          schema:
            example: {
              "newComp": "ABC",
              "email": "regularUser@email.com"
            }
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          schema:
            example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
    post:
      tags:
        - "Company"
      summary: "add a company"
      consumes:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "a new company"
          schema:
            example: {
              "companyName": "aCompany"
            }
      responses:
        200:
          description: "successful operation"
          schema:
            example: {
              "companyName": "aCompany"
            }
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          schema:
            example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"

    delete:
      tags:
        - "Company"
      summary: "delete a company"
      consumes:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "an existing company"
          schema:
            example: {
              "companyName": "Company2Delete"
            }
      responses:
        200:
          description: "successful operation"
          schema:
            example: {
              "companyName": "Company2Delete"
            }
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          schema:
            example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
  /user/subscription:
    put:
      tags:
        - "User"
      summary: "change a user's subscription status"
      consumes:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "a user email and new status"
          schema:
            example: {
              "isSubscribed": true,
              "email": "regularUser@email.com"
            }
      responses:
        200:
          description: "successful operation"
          schema:
            example: {
              "isSubscribed": true,
              "email": "regularUser@email.com"
            }
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          schema:
            example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
  /companies:
    get:
      tags:
        - "Company"
      summary: "get all companies"


      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            example: {
              "allCompanies": [
                "Galvanize",
                "Hootsuite",
                "Later"
              ]
            }
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          schema:
            example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
  /updateSettings:
    put:
      tags:
        - "Settings"
      summary: "Update system settings"
      parameters:
      - in: "body"
        name: "body"
        description: "change system settings for both scheduler and permission"
        schema:
          $ref: "#/definitions/Setting"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          schema:
            example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
  /permission:
    get:
      tags:
        - "Permission"
      summary: "Get permission settings of all user types"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            example: {
              "permission": [
              {
                "TYPE": "admin",
                "THRESHOLD": 2
              },
              {
                "TYPE": "regular",
                "THRESHOLD": 27
              },
              {
                "TYPE": "sysManager",
                "THRESHOLD": 1
              }
              ]
            }
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          schema:
            example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
definitions:
...
  User:
    type: "object"
    required:
      - "USER_NAME"
      - "EMAIL"
      - "TYPE"
      - "companyName"
      - "SUBSCRIBED"
    properties:
      USER_NAME:
        example: "John Doe"
      EMAIL:
        example: "regularUser@gmail.com"
      TYPE:
        example: "regular"
      companyName:
        example: "Galvanize"
      SUBSCRIBED:
        example: 1
  ArrayOfUsers:
    type: "object"
    required:
      - "users"
    properties:
      users:
        example: [
        {
          "USER_NAME": "Jackson Doe",
          "EMAIL": "adminUser@email.com",
          "TYPE": "admin",
          "companyName": null,
          "SUBSCRIBED": 1
        },
        {
          "USER_NAME": "Hee Su Kim",
          "EMAIL": "kessris@gmail.com",
          "TYPE": "sysManager",
          "companyName": null,
          "SUBSCRIBED": 1
        }]
...
  Setting:
    type: "object"
    required:
      - "permission"
      - "scheduler"
    properties:
      permission:
        example: [{"newThreshold": 27, "permissionType": "regular"},{"newThreshold": 2, "permissionType": "admin"}]
      scheduler:
        example: [{"taskId": "approval-reminder", "newFreq": 3, "maxNum": 17}, {"taskId": "subscription-email", "newFreq": 1, "maxNum": 17}]
...